# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "description": "Documentation de l'API Laravel pour la gestion des produits, rayons, ventes, et utilisateurs.",
        "contact": {
            "email": "support@magana.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Serveur API principal"
        }
    ],
    "paths": {
...
        "/rayons": {
            "get": {
                "tags": [
                    "Rayons"
                ],
                "summary": "Liste de tous les rayons",
                "operationId": "799b3cf401d12ebda50b057a55822dac",
                "responses": {
                    "200": {
                        "description": "Liste des rayons"
                    }
                }
            },
            "post": {
                "tags": [
                    "Rayons"
                ],
                "summary": "Ajouter un nouveau rayon",
                "operationId": "383e200fefeffacc6fb1a0e08b888915",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rayon créé avec succès"
                    },
                    "400": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/rayons/{id}": {
            "get": {
                "tags": [
                    "Rayons"
                ],
                "summary": "Obtenir les détails d'un rayon",
                "operationId": "cb1d594518eba992e5cd75d337ca5419",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du rayon",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du rayon"
                    },
                    "404": {
                        "description": "Rayon non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Rayons"
                ],
                "summary": "Mettre à jour un rayon",
                "operationId": "83b25c2a0ead6da0bcc0376a008914fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du rayon",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rayon mis à jour"
                    },
                    "404": {
                        "description": "Rayon non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rayons"
                ],
                "summary": "Supprimer un rayon",
                "operationId": "1a23151c95ca23b64a6632034b06f023",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du rayon",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rayon supprimé"
                    },
                    "404": {
                        "description": "Rayon non trouvé"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}